Separated out some utility classes and methods to different files/modules/classes.  This was done so that the lookbackmultiple_calc.py module can run without dependencies on the PyQt or Qt libraries.
